Move LinkCollector to a new collector.py module.
