+++ tests/unit/test_finder.py
@@ -62,7 +62,7 @@ def test_no_partial_name_match(data):
 
 def test_tilde():
     """Finder can accept a path with ~ in it and will normalize it."""
-    with patch('pip._internal.index.os.path.exists', return_value=True):
+    with patch('pip._internal.collector.os.path.exists', return_value=True):
         finder = make_test_finder(find_links=['~/python-pkgs'])
     req = install_req_from_line("gmpy")
     with pytest.raises(DistributionNotFound):
