+++ tests/unit/test_index.py
@@ -1,10 +1,6 @@
 import logging
-import os.path
-from textwrap import dedent
 
 import pytest
-from mock import Mock, patch
-from pip._vendor import html5lib, requests
 from pip._vendor.packaging.specifiers import SpecifierSet
 
 from pip._internal.download import PipSession
@@ -12,108 +8,22 @@ from pip._internal.index import (
     CandidateEvaluator,
     CandidatePreferences,
     FormatControl,
-    HTMLPage,
     LinkCollector,
     LinkEvaluator,
     PackageFinder,
     _check_link_requires_python,
-    _clean_link,
-    _determine_base_url,
     _extract_version_from_fragment,
     _find_name_version_sep,
-    _get_html_page,
     filter_unallowed_hashes,
-    group_locations,
 )
 from pip._internal.models.candidate import InstallationCandidate
-from pip._internal.models.index import PyPI
 from pip._internal.models.link import Link
 from pip._internal.models.search_scope import SearchScope
 from pip._internal.models.selection_prefs import SelectionPreferences
 from pip._internal.models.target_python import TargetPython
 from pip._internal.pep425tags import get_supported
 from pip._internal.utils.hashes import Hashes
-from pip._internal.utils.typing import MYPY_CHECK_RUNNING
-from tests.lib import CURRENT_PY_VERSION_INFO, make_test_search_scope
-
-if MYPY_CHECK_RUNNING:
-    from typing import List, Optional
-
-
-def make_fake_html_page(url):
-    html = dedent(u"""\
-    <html><head><meta name="api-version" value="2" /></head>
-    <body>
-    <a href="/abc-1.0.tar.gz#md5=000000000">abc-1.0.tar.gz</a>
-    </body></html>
-    """)
-    content = html.encode('utf-8')
-    headers = {}
-    return HTMLPage(content, url=url, headers=headers)
-
-
-def make_test_link_collector(
-    find_links=None,  # type: Optional[List[str]]
-):
-    # type: (...) -> LinkCollector
-    """
-    Create a LinkCollector object for testing purposes.
-    """
-    session = PipSession()
-    search_scope = make_test_search_scope(
-        find_links=find_links,
-        index_urls=[PyPI.simple_url],
-    )
-
-    return LinkCollector(
-        session=session,
-        search_scope=search_scope,
-    )
-
-
-def check_links_include(links, names):
-    """
-    Assert that the given list of Link objects includes, for each of the
-    given names, a link whose URL has a base name matching that name.
-    """
-    for name in names:
-        assert any(link.url.endswith(name) for link in links), (
-            'name {!r} not among links: {}'.format(name, links)
-        )
-
-
-class TestLinkCollector(object):
-
-    @patch('pip._internal.index._get_html_response')
-    def test_collect_links(self, mock_get_html_response, data):
-        expected_url = 'https://pypi.org/simple/twine/'
-
-        fake_page = make_fake_html_page(expected_url)
-        mock_get_html_response.return_value = fake_page
-
-        link_collector = make_test_link_collector(
-            find_links=[data.find_links]
-        )
-        actual = link_collector.collect_links('twine')
-
-        mock_get_html_response.assert_called_once_with(
-            expected_url, session=link_collector.session,
-        )
-
-        # Spot-check the CollectedLinks return value.
-        assert len(actual.files) > 20
-        check_links_include(actual.files, names=['simple-1.0.tar.gz'])
-
-        assert len(actual.find_links) == 1
-        check_links_include(actual.find_links, names=['packages'])
-
-        actual_pages = actual.pages
-        assert list(actual_pages) == [expected_url]
-        actual_page_links = actual_pages[expected_url]
-        assert len(actual_page_links) == 1
-        assert actual_page_links[0].url == (
-            'https://pypi.org/abc-1.0.tar.gz#md5=000000000'
-        )
+from tests.lib import CURRENT_PY_VERSION_INFO
 
 
 def make_mock_candidate(version, yanked_reason=None, hex_digest=None):
@@ -856,61 +766,6 @@ class TestPackageFinder:
         assert evaluator._supported_tags == [('py36', 'none', 'any')]
 
 
-def test_group_locations__file_expand_dir(data):
-    """
-    Test that a file:// dir gets listdir run with expand_dir
-    """
-    files, urls = group_locations([data.find_links], expand_dir=True)
-    assert files and not urls, (
-        "files and not urls should have been found at find-links url: %s" %
-        data.find_links
-    )
-
-
-def test_group_locations__file_not_find_link(data):
-    """
-    Test that a file:// url dir that's not a find-link, doesn't get a listdir
-    run
-    """
-    files, urls = group_locations([data.index_url("empty_with_pkg")])
-    assert urls and not files, "urls, but not files should have been found"
-
-
-def test_group_locations__non_existing_path():
-    """
-    Test that a non-existing path is ignored.
-    """
-    files, urls = group_locations([os.path.join('this', 'doesnt', 'exist')])
-    assert not urls and not files, "nothing should have been found"
-
-
-@pytest.mark.parametrize(
-    ("html", "url", "expected"),
-    [
-        (b"<html></html>", "https://example.com/", "https://example.com/"),
-        (
-            b"<html><head>"
-            b"<base href=\"https://foo.example.com/\">"
-            b"</head></html>",
-            "https://example.com/",
-            "https://foo.example.com/",
-        ),
-        (
-            b"<html><head>"
-            b"<base><base href=\"https://foo.example.com/\">"
-            b"</head></html>",
-            "https://example.com/",
-            "https://foo.example.com/",
-        ),
-    ],
-)
-def test_determine_base_url(html, url, expected):
-    document = html5lib.parse(
-        html, transport_encoding=None, namespaceHTMLElements=False,
-    )
-    assert _determine_base_url(document, url) == expected
-
-
 @pytest.mark.parametrize(
     ("fragment", "canonical_name", "expected"),
     [
@@ -999,115 +854,3 @@ def test_find_name_version_sep_failure(fragment, canonical_name):
 def test_extract_version_from_fragment(fragment, canonical_name, expected):
     version = _extract_version_from_fragment(fragment, canonical_name)
     assert version == expected
-
-
-def test_request_http_error(caplog):
-    caplog.set_level(logging.DEBUG)
-    link = Link('http://localhost')
-    session = Mock(PipSession)
-    session.get.return_value = resp = Mock()
-    resp.raise_for_status.side_effect = requests.HTTPError('Http error')
-    assert _get_html_page(link, session=session) is None
-    assert (
-        'Could not fetch URL http://localhost: Http error - skipping'
-        in caplog.text
-    )
-
-
-def test_request_retries(caplog):
-    caplog.set_level(logging.DEBUG)
-    link = Link('http://localhost')
-    session = Mock(PipSession)
-    session.get.side_effect = requests.exceptions.RetryError('Retry error')
-    assert _get_html_page(link, session=session) is None
-    assert (
-        'Could not fetch URL http://localhost: Retry error - skipping'
-        in caplog.text
-    )
-
-
-@pytest.mark.parametrize(
-    ("url", "clean_url"),
-    [
-        # URL with hostname and port. Port separator should not be quoted.
-        ("https://localhost.localdomain:8181/path/with space/",
-         "https://localhost.localdomain:8181/path/with%20space/"),
-        # URL that is already properly quoted. The quoting `%`
-        # characters should not be quoted again.
-        ("https://localhost.localdomain:8181/path/with%20quoted%20space/",
-         "https://localhost.localdomain:8181/path/with%20quoted%20space/"),
-        # URL with IPv4 address and port.
-        ("https://127.0.0.1:8181/path/with space/",
-         "https://127.0.0.1:8181/path/with%20space/"),
-        # URL with IPv6 address and port. The `[]` brackets around the
-        # IPv6 address should not be quoted.
-        ("https://[fd00:0:0:236::100]:8181/path/with space/",
-         "https://[fd00:0:0:236::100]:8181/path/with%20space/"),
-        # URL with query. The leading `?` should not be quoted.
-        ("https://localhost.localdomain:8181/path/with/query?request=test",
-         "https://localhost.localdomain:8181/path/with/query?request=test"),
-        # URL with colon in the path portion.
-        ("https://localhost.localdomain:8181/path:/with:/colon",
-         "https://localhost.localdomain:8181/path%3A/with%3A/colon"),
-        # URL with something that looks like a drive letter, but is
-        # not. The `:` should be quoted.
-        ("https://localhost.localdomain/T:/path/",
-         "https://localhost.localdomain/T%3A/path/"),
-        # VCS URL containing revision string.
-        ("git+ssh://example.com/path to/repo.git@1.0#egg=my-package-1.0",
-         "git+ssh://example.com/path%20to/repo.git@1.0#egg=my-package-1.0"),
-        # URL with Windows drive letter. The `:` after the drive
-        # letter should not be quoted. The trailing `/` should be
-        # removed.
-        pytest.param(
-            "file:///T:/path/with spaces/",
-            "file:///T:/path/with%20spaces",
-            marks=pytest.mark.skipif("sys.platform != 'win32'"),
-        ),
-        # URL with Windows drive letter, running on non-windows
-        # platform. The `:` after the drive should be quoted.
-        pytest.param(
-            "file:///T:/path/with spaces/",
-            "file:///T%3A/path/with%20spaces/",
-            marks=pytest.mark.skipif("sys.platform == 'win32'"),
-        ),
-    ]
-)
-def test_clean_link(url, clean_url):
-    assert(_clean_link(url) == clean_url)
-
-
-class TestHTMLPage:
-
-    @pytest.mark.parametrize(
-        ('anchor_html, expected'),
-        [
-            # Test not present.
-            ('<a href="/pkg1-1.0.tar.gz"></a>', None),
-            # Test present with no value.
-            ('<a href="/pkg2-1.0.tar.gz" data-yanked></a>', ''),
-            # Test the empty string.
-            ('<a href="/pkg3-1.0.tar.gz" data-yanked=""></a>', ''),
-            # Test a non-empty string.
-            ('<a href="/pkg4-1.0.tar.gz" data-yanked="error"></a>', 'error'),
-            # Test a value with an escaped character.
-            ('<a href="/pkg4-1.0.tar.gz" data-yanked="version &lt 1"></a>',
-                'version < 1'),
-            # Test a yanked reason with a non-ascii character.
-            (u'<a href="/pkg-1.0.tar.gz" data-yanked="curlyquote \u2018"></a>',
-                u'curlyquote \u2018'),
-        ]
-    )
-    def test_iter_links__yanked_reason(self, anchor_html, expected):
-        html = (
-            # Mark this as a unicode string for Python 2 since anchor_html
-            # can contain non-ascii.
-            u'<html><head><meta charset="utf-8"><head>'
-            '<body>{}</body></html>'
-        ).format(anchor_html)
-        html_bytes = html.encode('utf-8')
-        page = HTMLPage(html_bytes, url='https://example.com/simple/')
-        links = list(page.iter_links())
-        link, = links
-        actual = link.yanked_reason
-        assert actual == expected
